+++ build.rs
@@ -1,8 +1,12 @@
 fn main() {
-    println!("cargo:rerun-if-changed=c_code/src/test_sve.c");
-    println!("cargo:rerun-if-changed=c_code/src/test_sve.h");
+    println!("cargo:rerun-if-changed=sve/src/sve_inv_sbox.c");
+    println!("cargo:rerun-if-changed=sve/src/sve_inv_sbox.h");
+    println!("cargo:rerun-if-changed=sve/src/inv_sbox.h");
+    println!("cargo:rerun-if-changed=sve/src/inv_sbox.h");
+
     cc::Build::new()
-        .file("c_code/src/test_sve.c")
+        .file("sve/src/sve_inv_sbox.c")
+        .file("sve/src/inv_sbox.c")
         .flag("-march=armv8-a+sve")
         .compile("sve");
 }
