+++ sve/src/inv_sbox.h
@@ -1,0 +1,1 @@
+#define STATE_WIDTH 12
